voice_email_GUI/requirements.txt
python-dotenv==1.0.1
SpeechRecognition==3.10.4
PyAudio==0.2.14
pyttsx3==2.90
